+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.3 ] - 2024/02/28
+
+  * Updated service `get_show_episodes` to include the `limit_total` argument.
+  * Added service `get_player_queue_info` to retrieve the player queue information.
+  * Added service `get_player_devices` to retrieve player device list.
+
 ###### [ 1.0.2 ] - 2024/02/28
 
   * Updated underlying `spotifywebapiPython` package requirement to version 1.0.31.